Add files via upload
Recreating the findings of http://econ.geo.uu.nl/peeg/peeg1829.pdf paper for patent and GDP trends of Metropolitan Statistical Areas. The patent data is based on 376 combined MPAs over 6 years of data (between 2010 and 2015). The GDP data is based on 381 MPAs between 2011 and 2016. The multiple years used to analyze the GDP data is likely the reason the GDP scaling versus population is slightly sublinear. Patents are scaling superlinearly against population and have yearly coefficient of determinations below 0.5. GDP on the other hand has an extremely high coefficient of determination when compared with population. The yearly totals of Patents and GDP are strongly linear and have correlation above 0.95.