Add files via upload
Analysis of Filipino Family Income Data relating to income and expenditure. The primary goal of the project was classification through various machine learning methods. The models used a variety of continuous and categorical predictors to predict whether an individual family's medical expenditure would in the top ten percent compared to the other families.